+++ go.mod
@@ -52,7 +52,7 @@ require (
 	github.com/tendermint/starport v0.19.5
 	github.com/tendermint/tendermint v0.34.19
 	github.com/tendermint/tm-db v0.6.7
-	go.buf.build/grpc/go/sonr-io/motor v1.4.27
+	go.buf.build/grpc/go/sonr-io/motor v1.4.28
 	golang.org/x/crypto v0.0.0-20220525230936-793ad666bf5e
 	golang.org/x/mobile v0.0.0-20220518205345-8578da9835fd
 	golang.org/x/net v0.0.0-20220624214902-1bab6f366d9e // indirect
